requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
todoist_api_python~=2.1.7
tqdm~=4.67.1